requirements/base.txt
Django==4.0.5
psycopg2==2.9.3
django-extensions==3.1.5
django-environ==0.9.0
whitenoise==6.2.0
celery==5.2.7
django-celery-results==2.4.0
django-taggit==3.0.0